refactor(test): adjust import paths for lifecycle tests
