+++ test/main/lifecycle/databaseInitHook.test.ts
@@ -2,7 +2,8 @@ import { describe, it, expect, beforeEach, vi } from 'vitest'
 import { databaseInitHook } from '../../../src/main/lib/lifecycle/hooks/databaseInitHook'
 import { DatabaseInitializer } from '../../../src/main/lib/lifecycle/DatabaseInitializer'
 import { getInstance } from '../../../src/main/presenter'
-import { LifecycleContext, LifecyclePhase } from '../../../src/shared/presenter'
+import { LifecycleContext } from './../../../src/shared/presenter'
+import { LifecyclePhase } from './../../../src/shared/lifecycle'
 
 // Mock dependencies
 vi.mock('../../src/main/lib/lifecycle/DatabaseInitializer')
